gemini/requirements_gemini.txt
# Additional requirements for Google Gemini integration
# Add these to your main requirements.txt

# Google Generative AI integration for LangChain
langchain-google-genai==1.0.10

# Google AI Python SDK (dependency for langchain-google-genai)
google-generativeai>=0.5.0

# Core LangChain and LangGraph dependencies (if not already installed)
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.1.0

# Streamlit for web interface (if not already installed)
streamlit>=1.28.0
